+++ config/default/theme.ts
@@ -1,0 +1,3 @@
+export default {
+  colorPrimary: '#1677FF'
+} as ProAliasToken