+++ config/index.ts
@@ -1,9 +1,11 @@
+import theme from './default/theme'
 import network from './default/network'
 import defaultSettings from './default/defaultSettings'
 
 export * from './default/proxy'
 
 export {
+  theme,
   network,
   defaultSettings
 }