+++ src/store/modules/global.ts
@@ -1,8 +1,16 @@
 import { reactive, toRefs } from 'vue'
 import { defineStore } from 'pinia'
-import type { BasicLayoutProps } from '@gx-design-vue/pro-layout'
 import type { ThemeConfig } from '@gx-design-vue/pro-provider'
 import { themeConfig as proThemeConfig } from '@gx-design-vue/pro-provider'
+import type { BasicLayoutProps } from '@gx-design-vue/pro-layout'
+import { handleThemeConfig } from '@gx-design-vue/pro-layout'
+import { defaultSettings, theme } from '@gx-config'
+import logo from '@/assets/logo.png'
+
+const layoutThemeConfig = {
+  ...proThemeConfig,
+  title: defaultSettings.title,
+} as ThemeConfig
 
 /**
  * @Author      gx12358
@@ -11,8 +19,12 @@ import { themeConfig as proThemeConfig } from '@gx-design-vue/pro-provider'
  * @description store-global 全局属性
  */
 export interface GlobalState {
-  settings: Partial<BasicLayoutProps & { showProgressBar: boolean; keepAlive: boolean; disabledScrollTop: boolean; }>;
-  animateSetting: ThemeConfig['animate']
+  isLogut: boolean;
+  settings: {
+    layout: BasicLayoutProps;
+    showProgressBar: boolean;
+    keepAlive: boolean;
+  };
 }
 
 type GlobalStateKey = keyof GlobalState
@@ -20,11 +32,14 @@ type GlobalStateKey = keyof GlobalState
 export const useStoreGlobal = defineStore('global', () => {
   const state = reactive<GlobalState>({
     settings: {
-      disabledScrollTop: false,
+      layout: {
+        logo,
+        ...handleThemeConfig({ ...layoutThemeConfig, primaryColor: theme.colorPrimary })
+      },
       keepAlive: true,
-      showProgressBar: true,
+      showProgressBar: true
     },
-    animateSetting: proThemeConfig.animate as ThemeConfig['animate']
+    isLogut: false
   })
 
   const setGlobalData: (params: Partial<Record<GlobalStateKey, GlobalState[GlobalStateKey]>>) => void = (params) => {